triangle example now uses "Licht"
