+++ Examples/TriangleExample/TriangleExample.csproj
@@ -9,30 +9,34 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\Catalyze\Catalyze.csproj" />
+        <ProjectReference Include="..\..\Licht.Vulkan\Licht.Vulkan.csproj" />
+        <ProjectReference Include="..\..\Licht\Licht.csproj" />
     </ItemGroup>
 
+    <!--Shader compilation-->
     <ItemGroup>
-        <ShaderDir Include="$(ProjectDir)\Shaders\**" />
-        <ShaderFiles Include="$(ProjectDir)**\*.spv" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Update="@(ShaderFiles)">
-        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </None>
+        <None Update="$(ProjectDir)\assets\**\*">
+            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </None>
+        <ShaderDir Include="$(ProjectDir)\assets\shaders\*" />
+        <CompiledShaders Include="$(ProjectDir)\assets\shaders\**\*.spv" />
     </ItemGroup>
 
     <Target Name="CleanCompiledShaders" AfterTargets="Clean">
-        <Message Text="Clean Compiled Shaders \n@(ShaderFiles)" />
-        <Delete Files="@(ShaderFiles)" />
+        <Message Text="Clean compiled shaders \n@(CompiledShaders)" />
+        <Delete Files="@(CompiledShaders)" />
     </Target>
 
     <Target Name="CompileShaders" BeforeTargets="ResolveReferences">
-        <Message Text="Compile Shaders \n@(ShaderDir)@(VertexShader)" />
+        <Message Text="Compile Shaders \n@(ShaderDir)" />
         <Exec Command="glslc &quot;%(ShaderDir.FullPath)&quot; -o &quot;%(ShaderDir.FullPath).spv&quot;" Condition="'%(ShaderDir.Extension)' != '.spv'" />
+        <Message Text="Copy Shaders \n@(CompiledShaders)" />
         <ItemGroup>
-            <EmbeddedResource Include="$(ProjectDir)**\*.spv" />
+            <None Include="@(CompiledShaders)">
+                <Link>assets/shaders/%(Filename)%(Extension)</Link>
+                <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+            </None>
         </ItemGroup>
     </Target>
+    <!--/Shader compilation-->
 </Project>
