+++ Examples/TriangleExample/assets/shaders/triangle.frag
