+++ Examples/TriangleExample/assets/shaders/triangle.vert.spv
