+++ Licht.sln
@@ -15,8 +15,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ComputeShader", "Examples\C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Licht", "Licht\Licht.csproj", "{BA18A08A-E169-466A-B83F-C470A7B3B00F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Temp", "Temp\Temp.csproj", "{BBDB9BE8-F99B-478A-98DF-E3FFB2CBB100}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Licht.GraphicsCore", "Licht.GraphicsCore\Licht.GraphicsCore.csproj", "{9C15FB6A-8421-4351-97EA-1B0B9CEFB6B7}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Licht.Vulkan", "Licht.Vulkan\Licht.Vulkan.csproj", "{54BBC41E-7E79-4C5D-92A9-0D3741B37BCC}"
@@ -49,10 +47,6 @@ Global
 		{BA18A08A-E169-466A-B83F-C470A7B3B00F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BA18A08A-E169-466A-B83F-C470A7B3B00F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BA18A08A-E169-466A-B83F-C470A7B3B00F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BBDB9BE8-F99B-478A-98DF-E3FFB2CBB100}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BBDB9BE8-F99B-478A-98DF-E3FFB2CBB100}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BBDB9BE8-F99B-478A-98DF-E3FFB2CBB100}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BBDB9BE8-F99B-478A-98DF-E3FFB2CBB100}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9C15FB6A-8421-4351-97EA-1B0B9CEFB6B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9C15FB6A-8421-4351-97EA-1B0B9CEFB6B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9C15FB6A-8421-4351-97EA-1B0B9CEFB6B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
