--- Temp/Temp.csproj
@@ -1,41 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Licht.Vulkan\Licht.Vulkan.csproj" />
-      <ProjectReference Include="..\Licht\Licht.csproj" />
-    </ItemGroup>
-
-    <!--Shader compilation-->
-    <ItemGroup>
-        <None Update="$(ProjectDir)\assets\**\*">
-            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-        </None>
-        <ShaderDir Include="$(ProjectDir)\assets\shaders\*" />
-        <CompiledShaders Include="$(ProjectDir)\assets\shaders\**\*.spv" />
-    </ItemGroup>
-
-    <Target Name="CleanCompiledShaders" AfterTargets="Clean">
-        <Message Text="Clean compiled shaders \n@(CompiledShaders)" />
-        <Delete Files="@(CompiledShaders)" />
-    </Target>
-
-    <Target Name="CompileShaders" BeforeTargets="ResolveReferences">
-        <Message Text="Compile Shaders \n@(ShaderDir)" />
-        <Exec Command="glslc &quot;%(ShaderDir.FullPath)&quot; -o &quot;%(ShaderDir.FullPath).spv&quot;" Condition="'%(ShaderDir.Extension)' != '.spv'" />
-        <Message Text="Copy Shaders \n@(CompiledShaders)" />
-        <ItemGroup>
-            <None Include="@(CompiledShaders)">
-                <Link>assets/shaders/%(Filename)%(Extension)</Link>
-                <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-            </None>
-        </ItemGroup>
-    </Target>
-    <!--/Shader compilation-->
-</Project>
