--- Temp/assets/shaders/triangle.frag
@@ -1,8 +1,0 @@
-#version 450
-
-layout (location = 0) out vec4 outFragColor;
-
-void main()
-{
-    outFragColor = vec4(1.f,0.f,0.f,1.0f);
-}
